+++ libs/aws/sabi-s3/src/core/mod.rs
@@ -5,10 +5,5 @@ mod error;
 pub use error::Error;
 pub use error::Result;
 
-mod e_tag;
-pub use e_tag::ETag;
-
-pub mod headers;
-pub use headers::S3HeaderMap;
-
+pub mod response;
 pub mod verbs;
