+++ libs/aws/sabi-s3/src/core/response/headers/e_tag.rs
