+++ libs/aws/sabi-s3/src/core/response/headers/error.rs
